Move the class to the formatting package and the rule to the formatting ruleset
