+++ src/main/groovy/org/codenarc/rule/formatting/ClassEndsWithBlankLineBeforeClosingBraceRequiredRule.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.codenarc.rule.braces
+package org.codenarc.rule.formatting
 
 import groovy.transform.Memoized
 import org.codehaus.groovy.ast.ASTNode
