+++ src/main/resources/rulesets/formatting.xml
@@ -39,4 +39,5 @@
     <rule class='org.codenarc.rule.formatting.SpaceBeforeClosingBraceRule'/>
     <rule class='org.codenarc.rule.formatting.SpaceBeforeOpeningBraceRule'/>
     <rule class='org.codenarc.rule.formatting.TrailingWhitespaceRule'/>
+    <rule class='org.codenarc.rule.formatting.ClassEndsWithBlankLineBeforeClosingBraceRequiredRule'/>
 </ruleset>
