+++ src/test/resources/RunCodeNarcAgainstProjectSourceCode.ruleset
@@ -7,9 +7,7 @@ ruleset {
 
     ruleset('rulesets/basic.xml')
 
-    ruleset('rulesets/braces.xml') {
-        exclude 'ClassEndsWithBlankLineBeforeClosingBraceRequired'
-    }
+    ruleset('rulesets/braces.xml')
 
     ruleset('rulesets/convention.xml') {
         exclude 'NoDef'
@@ -44,6 +42,7 @@ ruleset {
         exclude 'SpaceAfterSwitch'
         exclude 'SpaceAfterWhile'
         exclude 'SpaceAroundMapEntryColon'
+        exclude 'ClassEndsWithBlankLineBeforeClosingBraceRequired'
         TrailingWhitespace(doNotApplyToFileNames:'ClassEndsWithBlankLineBeforeClosingBraceRequiredRuleTest.groovy') //Test should contain whitespaces to test the rule
     }
